requirements.txt
flask
python-dotenv
chromadb
sentence-transformers
groq
PyPDF2
CD